python/refloom_worker/pyproject.toml
[project]
name = "refloom-worker"
version = "0.1.0"
requires-python = ">=3.12"
dependencies = [
    "PyMuPDF>=1.25",
    "ebooklib>=0.18",
    "beautifulsoup4>=4.13",
    "pyobjc-framework-Vision>=11.0; sys_platform == 'darwin'",
    "pyobjc-framework-Quartz>=11.0; sys_platform == 'darwin'",
]

[tool.pytest.ini_options]
testpaths = ["."]

[tool.ruff.lint]
select = ["E", "F", "I", "W", "UP", "B", "SIM"]
ignore = ["SIM108"]

[tool.ruff]
target-version = "py312"
line-length = 120

[tool.pyright]
pythonVersion = "3.12"
typeCheckingMode = "basic"
venvPath = "."
venv = ".venv"
reportMissingImports = "warning"
reportAttributeAccessIssue = "warning"

[dependency-groups]
dev = [
    "pyright>=1.1.408",
    "pytest>=8.0",
    "ruff>=0.15.7",
]
